requirements.txt
--extra-index-url https://test.pypi.org/simple/
python-dateutil==2.8.1
flask==1.1.2
gunicorn===20.0.4
odd-contract-dev==0.0.22
more-itertools==8.6.0
pytz==2021.1
kfp==1.4.0
